backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
langgraph
langchain
langchain-openai
langchain-core
openai
python-dotenv